refactor: constants 파일 변경
