+++ src/domains/shared/constants/index.ts
@@ -22,3 +22,5 @@ export const Color = {
 };
 
 export type ColorType = keyof typeof Color;
+
+export const BASE_URL = process.env.NEXT_PUBLIC_API_URL;
